requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-tenants==3.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
uritemplate==4.2.0
zope.interface==7.2
